Extract tag stories to tag articles controller (#13457)
