+++ app/controllers/stories_controller.rb
@@ -27,7 +27,6 @@ class StoriesController < ApplicationController
     @article_index = true
 
     return handle_user_or_organization_or_podcast_or_page_index if params[:username]
-    return handle_tag_index if params[:tag]
 
     handle_base_index
   end
@@ -123,42 +122,6 @@ class StoriesController < ApplicationController
       Honeycomb.add_field("stories_route", "user")
       handle_user_index
     end
-  end
-
-  def handle_tag_index
-    @tag = params[:tag].downcase
-    @page = (params[:page] || 1).to_i
-    @tag_model = Tag.find_by(name: @tag) || not_found
-    @moderators = User.with_role(:tag_moderator, @tag_model).select(:username, :profile_image, :id)
-    if @tag_model.alias_for.present?
-      redirect_permanently_to("/t/#{@tag_model.alias_for}")
-      return
-    end
-
-    @num_published_articles = if @tag_model.requires_approval?
-                                @tag_model.articles.published.where(approved: true).count
-                              elsif Settings::UserExperience.feed_strategy == "basic"
-                                tagged_count
-                              else
-                                Rails.cache.fetch("article-cached-tagged-count-#{@tag}", expires_in: 2.hours) do
-                                  tagged_count
-                                end
-                              end
-    @number_of_articles = user_signed_in? ? 5 : SIGNED_OUT_RECORD_COUNT
-    @stories = Articles::Feeds::LargeForemExperimental
-      .new(number_of_articles: @number_of_articles, tag: @tag, page: @page)
-      .published_articles_by_tag
-
-    @stories = @stories.where(approved: true) if @tag_model&.requires_approval
-
-    @stories = stories_by_timeframe
-    @stories = @stories.decorate
-
-    set_surrogate_key_header "articles-#{@tag}"
-    set_cache_control_headers(600,
-                              stale_while_revalidate: 30,
-                              stale_if_error: 86_400)
-    render template: "articles/tag_index"
   end
 
   def handle_page_display
@@ -348,17 +311,6 @@ class StoriesController < ApplicationController
     @github_repositories = @user.github_repos.featured.order(stargazers_count: :desc, name: :asc)
   end
 
-  def stories_by_timeframe
-    if %w[week month year infinity].include?(params[:timeframe])
-      @stories.where("published_at > ?", Timeframe.datetime(params[:timeframe]))
-        .order(public_reactions_count: :desc)
-    elsif params[:timeframe] == "latest"
-      @stories.where("score > ?", -20).order(published_at: :desc)
-    else
-      @stories.order(hotness_score: :desc).where("score >= ?", Settings::UserExperience.home_feed_minimum_score)
-    end
-  end
-
   def assign_podcasts
     return unless user_signed_in?
 
@@ -484,9 +436,5 @@ class StoriesController < ApplicationController
     return :relevance unless params[:sort_by] == "published_at" && params[:sort_direction].present?
 
     params[:sort_direction] == "desc" ? :newest : :oldest
-  end
-
-  def tagged_count
-    @tag_model.articles.published.where("score >= ?", Settings::UserExperience.tag_feed_minimum_score).count
   end
 end
